requirements.txt
https://github.com/rambo/hachoir3/archive/3.0a2-rambo.zip
construct==2.8.12
python-dateutil==2.6.1
